1. change input arg from + to * 2. drop edge before drop vertex 3. merge and drop GdbDataParallelRemove.java 4. update README and CHANGELOG
